Migrated `surf-transaction` from Velocity to Paper. Removed Velocity-specific commands and arguments. Introduced Paper-specific command arguments, commands, and plugin structures. Updated gradle configuration and module settings accordingly. Refactored `TransactionRepository` and improved transaction result handling. Implemented `PaperMain` as the entry point for Paper platform.
