+++ surf-transaction-paper/src/main/kotlin/dev/slne/surf/transaction/paper/commands/currency/CurrencyCommand.kt
@@ -1,0 +1,10 @@
+package dev.slne.surf.transaction.paper.commands.currency
+
+import dev.jorel.commandapi.kotlindsl.commandTree
+import dev.slne.surf.transaction.paper.commands.CommandPermission
+import dev.slne.surf.transaction.paper.commands.currency.admin.currencyAdminCommand
+
+fun currencyCommand() = commandTree("currency") {
+    withPermission(CommandPermission.CURRENCY)
+    currencyAdminCommand()
+}
