+++ surf-transaction-paper/src/main/kotlin/dev/slne/surf/transaction/paper/commands/currency/admin/CurrencyAdminCommand.kt
@@ -1,0 +1,13 @@
+package dev.slne.surf.transaction.paper.commands.currency.admin
+
+import dev.jorel.commandapi.CommandTree
+import dev.jorel.commandapi.kotlindsl.literalArgument
+import dev.slne.surf.transaction.paper.commands.CommandPermission
+import dev.slne.surf.transaction.paper.commands.currency.admin.subcommands.currencyCreateCommand
+import dev.slne.surf.transaction.paper.commands.currency.admin.subcommands.currencyMakeDefaultCommand
+
+fun CommandTree.currencyAdminCommand() = literalArgument("admin") {
+    withPermission(CommandPermission.CURRENCY_ADMIN)
+    currencyCreateCommand()
+    currencyMakeDefaultCommand()
+}
