+++ surf-transaction-velocity/src/main/kotlin/dev/slne/surf/transaction/velocity/VelocityMain.kt
@@ -8,12 +8,7 @@ import com.velocitypowered.api.proxy.ProxyServer
 import dev.slne.surf.cloud.api.common.CloudInstance
 import dev.slne.surf.cloud.api.common.startSpringApplication
 import dev.slne.surf.transaction.SurfTransactionSpringApplication
-import dev.slne.surf.transaction.core.netty.packets.ClientboundRefreshCurrencies
 import dev.slne.surf.transaction.core.transactionApiBridgeImpl
-import dev.slne.surf.transaction.velocity.commands.balance.balanceCommand
-import dev.slne.surf.transaction.velocity.commands.currency.currencyCommand
-import dev.slne.surf.transaction.velocity.commands.pay.payCommand
-import dev.slne.surf.transaction.velocity.commands.transaction.transactionCommand
 import java.nio.file.Path
 
 lateinit var plugin: VelocityMain
@@ -28,16 +23,7 @@ class VelocityMain @Inject constructor(
     init {
         plugin = this
         suspendingPluginContainer.initialize(this)
-
-        println("org/springframework/scheduling/annotation/ProxyAsyncConfiguration: " + Class.forName("org.springframework.scheduling.annotation.ProxyAsyncConfiguration"))
-        println("ClientboundRefreshCurrencies serializer: " + ClientboundRefreshCurrencies.serializer())
-
         transactionApiBridgeImpl.context =
             CloudInstance.startSpringApplication(SurfTransactionSpringApplication::class)
-
-        transactionCommand()
-        currencyCommand()
-        balanceCommand()
-        payCommand()
     }
 }
