--- surf-transaction-velocity/src/main/kotlin/dev/slne/surf/transaction/velocity/commands/transaction/TransactionCommand.kt
@@ -1,10 +1,0 @@
-package dev.slne.surf.transaction.velocity.commands.transaction
-
-import dev.jorel.commandapi.kotlindsl.commandAPICommand
-import dev.slne.surf.transaction.velocity.commands.transaction.admin.transactionAdminCommand
-
-
-fun transactionCommand() = commandAPICommand("transaction") {
-    withPermission("surf.transaction")
-    transactionAdminCommand()
-}
